requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
transformers==5.3.0
torch==2.10.0
accelerate==1.13.0
sentencepiece==0.2.1
peft>=0.5.0
safetensors>=0.7.0
requests